refactor(ui): move overview types into each action folder
