+++ ui/actions/overview/attack-surface/attack-surface.ts
@@ -3,7 +3,7 @@
 import { apiBaseUrl, getAuthHeaders } from "@/lib";
 import { handleApiResponse } from "@/lib/server-actions-helper";
 
-import { AttackSurfaceOverviewResponse } from "../types";
+import { AttackSurfaceOverviewResponse } from "./types";
 
 export const getAttackSurfaceOverview = async ({
   filters = {},
