+++ ui/actions/overview/attack-surface/index.ts
@@ -1,2 +1,3 @@
 export * from "./attack-surface";
 export * from "./attack-surface.adapter";
+export * from "./types";
