+++ ui/actions/overview/findings/index.ts
@@ -1,1 +1,2 @@
 export * from "./findings";
+export * from "./types";
