+++ ui/actions/overview/providers/index.ts
@@ -1,2 +1,3 @@
 export * from "./providers";
 export * from "./sankey.adapter";
+export * from "./types";
