+++ ui/actions/overview/regions/index.ts
@@ -1,2 +1,3 @@
 export * from "./regions";
 export * from "./threat-map.adapter";
+export * from "./types";
