+++ ui/actions/overview/severity-trends/index.ts
@@ -1,2 +1,3 @@
 export * from "./severity-trends";
 export * from "./severity-trends.adapter";
+export * from "./types";
