+++ ui/actions/overview/threat-score/index.ts
@@ -1,1 +1,2 @@
 export * from "./threat-score";
+export * from "./types";
