--- ui/actions/overview/types/common.ts
@@ -1,5 +1,0 @@
-// Common types shared across overview endpoints
-
-export interface OverviewResponseMeta {
-  version: string;
-}
