--- ui/actions/overview/types/index.ts
@@ -1,8 +1,0 @@
-export * from "./attack-surface";
-export * from "./common";
-export * from "./findings-severity";
-export * from "./providers";
-export * from "./regions";
-export * from "./services";
-export * from "./severity-trends";
-export * from "./threat-score";
